Django/requirements.txt
asgiref==3.11.1
attrs==26.1.0
Django==6.0.2
django-environ==0.13.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mysqlclient==2.2.8
packaging==26.2
PyJWT==2.12.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
